+++ modules/sor/sorV2/lib/poolsV3/weighted/weightedPool.ts
@@ -3,7 +3,7 @@ import { MAX_UINT256, SwapKind, Token, TokenAmount, WAD } from '@balancer/sdk';
 import { AddKind, RemoveKind, Vault, Weighted, WeightedState } from '@balancer-labs/balancer-maths';
 import { Chain } from '@prisma/client';
 
-import { PrismaPoolWithDynamic } from '../../../../../../prisma/prisma-types';
+import { PrismaPoolAndHookWithDynamic } from '../../../../../../prisma/prisma-types';
 import { GqlPoolType } from '../../../../../../schema';
 import { TokenPairData } from '../../../../../sources/contracts/v3/fetch-tokenpair-data';
 import { chainToIdMap } from '../../../../../network/network-config';
@@ -32,7 +32,7 @@ export class WeightedPoolV3 implements BasePoolV3 {
     private vault: Vault;
     private poolState: WeightedState;
 
-    static fromPrismaPool(pool: PrismaPoolWithDynamic): WeightedPoolV3 {
+    static fromPrismaPool(pool: PrismaPoolAndHookWithDynamic): WeightedPoolV3 {
         const poolTokens: WeightedPoolToken[] = [];
 
         if (!pool.dynamicData) {